Moved tb-treatment query to common queries (#132)
* Moved tb-treatment querie to common queries

* Re-organize CommonCohortQueries

* Organize JavaDoc

* Delete unused TbCohortQueries
