--- api/src/main/java/org/openmrs/module/eptsreports/reporting/library/cohorts/TbCohortQueries.java
@@ -1,46 +1,0 @@
-package org.openmrs.module.eptsreports.reporting.library.cohorts;
-
-import java.util.Date;
-import org.openmrs.Location;
-import org.openmrs.module.eptsreports.metadata.HivMetadata;
-import org.openmrs.module.eptsreports.reporting.library.queries.TbQueries;
-import org.openmrs.module.reporting.cohort.definition.CohortDefinition;
-import org.openmrs.module.reporting.cohort.definition.SqlCohortDefinition;
-import org.openmrs.module.reporting.evaluation.parameter.Parameter;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Component;
-
-@Component
-public class TbCohortQueries {
-
-  @Autowired private HivMetadata hivMetadata;
-
-  /**
-   * Get All Patients On TB Treatment
-   *
-   * @return
-   */
-  public CohortDefinition getPatientsOnTbTreatment() {
-    SqlCohortDefinition cd = new SqlCohortDefinition();
-    cd.setName("patientsOnTbTreatment");
-    cd.addParameter(new Parameter("startDate", "startDate", Date.class));
-    cd.addParameter(new Parameter("endDate", "endDate", Date.class));
-    cd.addParameter(new Parameter("location", "location", Location.class));
-
-    cd.setQuery(
-        TbQueries.getPatientsOnTbTreatmentQuery(
-            hivMetadata.getAdultoSeguimentoEncounterType().getEncounterTypeId(),
-            hivMetadata.getPediatriaSeguimentoEncounterType().getEncounterTypeId(),
-            hivMetadata.getTBDrugStartDateConcept().getConceptId(),
-            hivMetadata.getTBDrugEndDateConcept().getConceptId(),
-            hivMetadata.getTBProgram().getProgramId(),
-            hivMetadata.getPatientActiveOnTBProgramWorkflowState().getProgramWorkflowStateId(),
-            hivMetadata.getActiveTBConcept().getConceptId(),
-            hivMetadata.getYesConcept().getConceptId(),
-            hivMetadata.getTBTreatmentPlanConcept().getConceptId(),
-            hivMetadata.getStartDrugs().getConceptId(),
-            hivMetadata.getContinueRegimenConcept().getConceptId()));
-
-    return cd;
-  }
-}
